Move dateUtils.ts from utils/formatters/ to utils/ and delete formatters folder
Related to #1232
