+++ src/app/utils/formatUtils.ts
