+++ src/app/utils/validationUtils.ts
